// src/enums.ts
export enum Sentiments {
  NOT_CALCULATED = -1,
  NOT_USED = 0,
  VERY_NEGATIVE = 1,
  NEGATIVE = 2,
  NEUTRAL = 3,
  POSITIVE = 4,
  VERY_POSITIVE = 5,
  NO_SENTIMENTS = 6,
  MIXED = 7,
}

export enum SentimentBackgroundColor {
  VERY_NEGATIVE = '#ff768180',
  NEGATIVE = '#ffb8ba80',
  NEUTRAL = '#ffdebb80',
  POSITIVE = '#b8ffb280',
  VERY_POSITIVE = '#1ba75880',
}